refactor: Split up Lootrun in more manageable parts
