+++ common/src/main/java/com/wynntils/gui/screens/WynntilsLootrunsScreen.java
@@ -17,7 +17,7 @@ import com.wynntils.gui.widgets.BackButton;
 import com.wynntils.gui.widgets.LootrunButton;
 import com.wynntils.gui.widgets.PageSelectorButton;
 import com.wynntils.gui.widgets.ReloadButton;
-import com.wynntils.models.lootruns.LootrunModel;
+import com.wynntils.models.lootruns.LootrunInstance;
 import com.wynntils.models.lootruns.event.LootrunCacheRefreshEvent;
 import com.wynntils.utils.CommonColors;
 import com.wynntils.utils.StringUtils;
@@ -31,7 +31,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-public final class WynntilsLootrunsScreen extends WynntilsMenuListScreen<LootrunModel.LootrunInstance, LootrunButton> {
+public final class WynntilsLootrunsScreen extends WynntilsMenuListScreen<LootrunInstance, LootrunButton> {
     private WynntilsLootrunsScreen() {
         super(Component.translatable("screens.wynntils.lootruns.name"));
 
@@ -93,7 +93,7 @@ public final class WynntilsLootrunsScreen extends WynntilsMenuListScreen<Lootrun
         if (hovered instanceof LootrunButton lootrunButton) {
             List<Component> tooltipLines;
 
-            LootrunModel.LootrunInstance currentLootrun = Models.Lootrun.getCurrentLootrun();
+            LootrunInstance currentLootrun = Models.Lootrun.getCurrentLootrun();
             if (currentLootrun != null
                     && Objects.equals(lootrunButton.getLootrun().name(), currentLootrun.name())) {
                 tooltipLines = List.of(
@@ -160,7 +160,7 @@ public final class WynntilsLootrunsScreen extends WynntilsMenuListScreen<Lootrun
     }
 
     private void renderDescription(PoseStack poseStack) {
-        LootrunModel.LootrunInstance currentLootrun = Models.Lootrun.getCurrentLootrun();
+        LootrunInstance currentLootrun = Models.Lootrun.getCurrentLootrun();
         if (currentLootrun != null) {
             poseStack.pushPose();
             poseStack.translate(20, 80, 0);
