+++ common/src/main/java/com/wynntils/models/lootruns/type/ColoredPath.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright © Wynntils 2023.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.models.lootruns.type;
+
+import java.util.List;
+
+public record ColoredPath(List<ColoredPoint> points) {}
