+++ common/src/main/java/com/wynntils/models/lootruns/type/ColoredPoint.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright © Wynntils 2023.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.models.lootruns.type;
+
+import net.minecraft.world.phys.Vec3;
+
+public record ColoredPoint(Vec3 vec3, int color) {}
