+++ common/src/main/java/com/wynntils/models/lootruns/type/LootrunPath.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright © Wynntils 2023.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.models.lootruns.type;
+
+import java.util.List;
+import net.minecraft.world.phys.Vec3;
+
+public record LootrunPath(List<Vec3> points) {}
