+++ common/src/main/java/com/wynntils/models/lootruns/type/LootrunSaveResult.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright © Wynntils 2023.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.models.lootruns.type;
+
+public enum LootrunSaveResult {
+    SAVED,
+    ERROR_SAVING,
+    ERROR_ALREADY_EXISTS
+}
