+++ common/src/main/java/com/wynntils/models/lootruns/type/LootrunUndoResult.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright © Wynntils 2023.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.models.lootruns.type;
+
+public enum LootrunUndoResult {
+    SUCCESSFUL,
+    ERROR_STAND_NEAR_POINT,
+    ERROR_NOT_FAR_ENOUGH
+}
